+++ generators/app/tasks/appSetup/coreFiles/packageJsonScripts.js
@@ -1,8 +1,9 @@
 module.exports = function packgeJsonScripts() {
   const packageJson = this.fs.readJSON(this.destinationPath(this.projectName, 'package.json'));
   packageJson.scripts = packageJson.scripts || {};
-  packageJson.scripts.android = 'react-native run-android --variant=qaDebug';
-  packageJson.scripts.ios = 'react-native run-ios --scheme qa';
+  packageJson.scripts.start = 'npx react-native start';
+  packageJson.scripts.android = 'npx react-native run-android --variant=qaDebug';
+  packageJson.scripts.ios = 'npx react-native run-ios --scheme qa';
   packageJson.scripts.test = 'jest --passWithNoTests';
   packageJson.scripts.coverage = 'jest --coverage --passWithNoTests';
   packageJson.scripts.clean = 'rm -rf $TMPDIR/react-* && watchman watch-del-all && yarn cache clean';
@@ -20,9 +21,8 @@ module.exports = function packgeJsonScripts() {
   packageJson.scripts.lint = 'eslint src --ext .js,.ts,.jsx,.tsx';
   packageJson.scripts['lint-fix'] = 'eslint src --ext .js,.ts,.jsx,.tsx --fix';
   packageJson.scripts['lint-diff'] =
-    'git diff --name-only --cached --no-error-on-unmatched-pattern --relative | grep -E "\\.(ts|tsx|js|jsx)$" | xargs eslint';
+    'git diff --name-only --relative --diff-filter=ACM | grep -E "\\.(ts|tsx|js|jsx)$" | xargs eslint';
   packageJson.scripts.tsc = 'tsc';
-
   packageJson.husky = packageJson.husky || {};
   packageJson.husky.hooks = packageJson.husky.hooks || {};
   packageJson.husky.hooks['pre-commit'] = 'yarn run lint-diff && yarn run tsc';