+++ generators/app/tasks/appSetup/featuresFiles/crashlyticsFeatureFiles.js
@@ -2,7 +2,7 @@ function addConfigToAndroidFiles() {
   let buildGradleContent = this.fs.read(`${this.projectName}/android/build.gradle`);
   buildGradleContent = buildGradleContent.replace(
     '// NOTE: Do not place your application dependencies here; they belong',
-    "classpath 'com.google.firebase:firebase-crashlytics-gradle:2.0.0'\n\t\t// NOTE: Do not place your application dependencies here; they belong"
+    "classpath 'com.google.firebase:firebase-crashlytics-gradle:2.2.0'\n\t\t// NOTE: Do not place your application dependencies here; they belong"
   );
   this.fs.write(`${this.projectName}/android/build.gradle`, buildGradleContent);
 
@@ -10,10 +10,6 @@ function addConfigToAndroidFiles() {
   appBuildGradleContent = appBuildGradleContent.replace(
     "apply plugin: 'com.google.gms.google-services'",
     "apply plugin: 'com.google.gms.google-services'\napply plugin: 'com.google.firebase.crashlytics'"
-  );
-  appBuildGradleContent = appBuildGradleContent.replace(
-    'proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"',
-    'proguardFiles getDefaultProguardFile("proguard-android.txt"), "proguard-rules.pro"\n\t\t\tfirebaseCrashlytics {\n\t\t\t\tnativeSymbolUploadEnabled true\n\t\t\t}'
   );
   this.fs.write(`${this.projectName}/android/app/build.gradle`, appBuildGradleContent);
 }