+++ generators/app/tasks/installTasks/installPods.js
@@ -2,8 +2,8 @@ const runCommand = require('../runCommand');
 
 module.exports = function installPods() {
   return runCommand({
-    command: ['pod', ['install', '--repo-update'], { cwd: `${process.cwd()}/${this.projectName}/ios` }],
-    loadingMessage: 'Updating and installing Pods...',
+    command: ['pod', ['install'], { cwd: `${process.cwd()}/${this.projectName}/ios` }],
+    loadingMessage: 'Installing Pods...',
     successMessage: 'Pods ready!',
     failureMessage: 'Pod install failed. Turn verbose mode on for detailed logging',
     context: this.options